Added Revoke Functionality to Router
Added the ability for the router to remove a route from its forwarding
table. This is accomplished by converting the offending route to a range
of ip addresses, and then removing any networks that fall within this
range. Any removed networks are then examined to ensure that no part of
them must remain.